+++ packages/components/pie-modal/src/defs.ts
@@ -6,14 +6,17 @@ export interface ModalProps {
      * The text to display in the modal's heading.
      */
     heading: string;
+
     /**
      * The HTML heading tag to use for the modal's heading. Can be h1-h6.
      */
     headingLevel: typeof headingLevels[number];
+
     /**
      * When true, the modal will be open.
      */
     isOpen: boolean;
+
     /**
      * When set to `true`:
      *  1. The close button within the modal will be visible.
@@ -28,13 +31,19 @@ export interface ModalProps {
     isDismissible: boolean;
 
     /**
+     * This controls whether a *medium-sized* modal will cover the full width of the page when below the mid breakpoint.
+     */
+    isFullWidthBelowMid: boolean;
+
+    /**
+     * The selector for the element that you would like focus to be returned to when the modal is closed, e.g., #skipToMain
+     */
+    returnFocusAfterCloseSelector?: string;
+
+    /**
      * The size of the modal; this controls how wide it will appear on the page.
      */
     size: typeof sizes[number];
-    /**
-     * The width of the modal; this controls whether it will cover the full width of the page when below the mid breakpoint.
-     */
-    isFullWidthBelowMid: boolean;
 }
 
 /**