+++ packages/components/pie-modal/test/component/pie-modal.spec.ts
@@ -1,7 +1,15 @@
 import { test, expect } from '@sand4rt/experimental-ct-web';
+import { type Page } from '@playwright/test';
 import { PieIconButton } from '@justeattakeaway/pie-icon-button';
+import {
+    WebComponentTestWrapper,
+} from '@justeattakeaway/pie-webc-testing/src/helpers/components/web-component-test-wrapper/WebComponentTestWrapper.ts';
+import { renderTestPieModal } from '../helpers/index.ts';
+
 import { PieModal } from '@/index';
 import { headingLevels } from '@/defs';
+
+const closeButtonSelector = '[data-test-id="modal-close-button"]';
 
 // Mount any components that are used inside of pie-modal so that
 // they have been registered with the browser before the tests run.
@@ -38,7 +46,6 @@ headingLevels.forEach((headingLevel) => test(`should render the correct heading 
     // Arrange
     const props = {
         heading: 'Modal Header',
-        // assert type checking as we purposely provide incorrect value
         headingLevel,
     };
 
@@ -47,13 +54,12 @@ headingLevels.forEach((headingLevel) => test(`should render the correct heading 
     // @ts-ignore // Added this as we want to deliberately test with invalid headingLevel (which is an invalid type based on ModalProps)
     const component = await mount(PieModal, { props });
 
-    // h2 is the default / fallback value
     // Assert
     await expect(component.locator('h2.c-modal-heading')).toContainText(props.heading);
 }));
 
-test.describe('`Pie Modal is closed`', () => {
-    test.describe('when via the close button click', () => {
+test.describe('When modal is closed', () => {
+    test.describe('by clicking the close button', () => {
         test('should dispatch event `pie-modal-close`', async ({ mount, page }) => {
             // Arrange
             const messages: string[] = [];
@@ -70,15 +76,14 @@ test.describe('`Pie Modal is closed`', () => {
                 },
             );
 
-            // Act
-            await page.locator('[data-test-id="modal-close-button"]').click();
+            await page.locator(closeButtonSelector).click();
 
             // Assert
             expect(messages).toHaveLength(1);
         });
     });
 
-    test.describe('when via the backdrop click', () => {
+    test.describe('by clicking the backdrop', () => {
         test('should dispatch event `pie-modal-close`', async ({ mount, page }) => {
             // Arrange
             const messages: string[] = [];
@@ -101,11 +106,126 @@ test.describe('`Pie Modal is closed`', () => {
             expect(messages).toHaveLength(1);
         });
     });
+
+    test.describe('`returnFocusAfterCloseSelector` prop', () => {
+        test.describe('when given', () => {
+            test('should return focus to specified element', async ({ mount, page }) => {
+                // Arrange
+                const component = renderTestPieModal({
+                    returnFocusAfterCloseSelector: '#focus-me',
+                });
+
+                await mount(WebComponentTestWrapper, {
+                    props: {
+                        pageMode: true,
+                    },
+                    slots: {
+                        component,
+                        pageMarkup: `<div>
+                            <button id="default"></button>
+                            <button id="focus-me"></button>
+                            <button id="not-me"></button>
+                        </div>`,
+                    },
+                });
+
+                // Act
+                await page.locator(closeButtonSelector).click();
+
+                const focusedElement = await page.locator(':focus');
+                const focusedElementId = await focusedElement.getAttribute('id');
+
+                // Assert
+                expect(focusedElementId).toBe('focus-me');
+            });
+
+            test('should return focus to first matching element', async ({ page, mount }) => {
+                // Arrange
+                const component = renderTestPieModal({
+                    returnFocusAfterCloseSelector: '[data-test-id="focus-me"]',
+                });
+
+                await mount(WebComponentTestWrapper, {
+                    props: {
+                        pageMode: true,
+                    },
+                    slots: {
+                        component,
+                        pageMarkup: `<div>
+                            <button id="default"></button>
+                            <button data-test-id="focus-me" id="actual-focus"></button>
+                            <button data-test-id="focus-me"></button>
+                        </div>`,
+                    },
+                });
+
+                // Act
+                await page.locator(closeButtonSelector).click();
+
+                const focusedElement = await page.locator(':focus');
+                const focusedElementId = await focusedElement.getAttribute('id');
+
+                // Assert
+                expect(focusedElementId).toBe('actual-focus');
+            });
+        });
+
+        test.describe('when not given', () => {
+            [{
+                mechanism: 'close button',
+                modalCloseFunction: async (page : Page) => {
+                    await page.locator(closeButtonSelector).click();
+                },
+            }, {
+                mechanism: 'Esc key',
+                modalCloseFunction: async (page : Page) => {
+                    await page.keyboard.press('Escape');
+                },
+            }].forEach(({ mechanism, modalCloseFunction }) => {
+                test.describe(`and closed by the ${mechanism}`, () => {
+                    test('should return focus to the element that opens the modal', async ({ page, mount }) => {
+                        // Arrange
+                        const component = renderTestPieModal({ isOpen: false });
+
+                        await mount(WebComponentTestWrapper, {
+                            props: {
+                                pageMode: true,
+                            },
+                            slots: {
+                                component,
+                                pageMarkup: `<div>
+                                    <button id="not-me"></button>
+                                    <button data-test-id="open-modal" id="default"></button>
+                                </div>`,
+                            },
+                        });
+
+                        await page.evaluate(() => {
+                            // Set up a button which opens the modal when clicked
+                            document.querySelector('[data-test-id="open-modal"]')?.addEventListener('click', () => {
+                                document.querySelector('pie-modal')?.setAttribute('isOpen', 'true');
+                            });
+                        });
+
+                        // Act
+                        await page.locator('[data-test-id="open-modal"]').click();
+                        await modalCloseFunction(page);
+
+                        const focusedElement = await page.locator(':focus');
+                        const focusedElementId = await focusedElement.getAttribute('id');
+
+                        // Assert
+                        expect(focusedElementId).toBe('default');
+                    });
+                });
+            });
+        });
+    });
 });
 
-test.describe('Pie Modal `isDismissible`', () => {
-    test.describe('when `isDismissible` is `true`', () => {
-        test('should contain a close button so the modal can be dismissed', async ({ mount }) => {
+test.describe('`isDismissible` prop', () => {
+    test.describe('when `true`', () => {
+        test('should make the modal contain a close button', async ({ mount }) => {
             // Arrange
             const component = await mount(
                 PieModal,
@@ -118,7 +238,7 @@ test.describe('Pie Modal `isDismissible`', () => {
             );
 
             // Act & Assert
-            await expect(component.locator('[data-test-id="modal-close-button"]')).toBeVisible();
+            await expect(component.locator(closeButtonSelector)).toBeVisible();
         });
 
         test('should close the modal when the close button is clicked', async ({ mount }) => {
@@ -168,7 +288,7 @@ test.describe('Pie Modal `isDismissible`', () => {
             expect(styles.display).toBe('none');
         });
 
-        test('should close the modal when the ESC key is triggered', async ({ mount, page }) => {
+        test('should close the modal when the Escape key is pressed', async ({ mount, page }) => {
             // Arrange
             const component = await mount(
                 PieModal,
@@ -189,7 +309,7 @@ test.describe('Pie Modal `isDismissible`', () => {
     });
 
     test.describe('when `isDismissible` is `false`', () => {
-        test('should NOT contain a close button so the modal can not be dismissed', async ({ mount }) => {
+        test('should make the modal NOT contain a close button', async ({ mount }) => {
             // Arrange
             const component = await mount(
                 PieModal,
@@ -202,7 +322,7 @@ test.describe('Pie Modal `isDismissible`', () => {
             );
 
             // Act & Assert
-            await expect(component.locator('[data-test-id="modal-close-button"]')).not.toBeVisible();
+            await expect(component.locator(closeButtonSelector)).not.toBeVisible();
         });
 
         test('should NOT close the modal when the backdrop is clicked', async ({ mount, page }) => {
@@ -233,7 +353,7 @@ test.describe('Pie Modal `isDismissible`', () => {
             expect(styles.display).toBe('block');
         });
 
-        test('should NOT close the modal when the ESC key is triggered', async ({ mount, page }) => {
+        test('should NOT close the modal when the Escape key is pressed', async ({ mount, page }) => {
             // Arrange
             const component = await mount(
                 PieModal,