+++ packages/components/pie-webc-testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.1
+
+### Patch Changes
+
+- [Added] - Add returnFocusAfterCloseSelector prop ([#599](https://github.com/justeattakeaway/pie/pull/599)) by [@xander-marjoram](https://github.com/xander-marjoram)
+
+  [Changed] - Extract `renderTestPieModal` into test helper file so it can be used for component tests too
+  [Added] - Add README documentation for new props
+
 ## 0.1.0
 
 ### Minor Changes